Refactor auxiliary and tools
Move `death_vector` into auxiliary, remove `linear_combination` since it is no longer needed, rename `snap_PL` to `snap_pl`
